+++ examples/roles/fabfile.py
@@ -3,7 +3,7 @@ import functools
 import fabricio
 
 from fabric import api as fab, colors
-from fabricio import tasks, docker
+from fabricio import tasks, docker, utils
 from fabricio.misc import AvailableVagrantHosts
 
 fab.env.roledefs.update(
@@ -21,20 +21,15 @@ def vagrant():
     )
 
 
-@fab.task(name='monkey-patch')
-def monkey_patch():
-    """
-    apply monkey patch
-    """
-    # replace fabricio.run by fabricio.local to run all commands on localhost
-    fabricio.run = functools.partial(fabricio.local, capture=True)
-
-    # uncomment row below to disable file uploading (e.g. docker-compose.yml)
-    # fab.put = lambda *args, **kwargs: None
-
-
 @fabricio.infrastructure
-def localhost():
+def localhost(force_local=False):
+    if utils.strtobool(force_local):
+        # replace fabricio.run by fabricio.local to run all commands locally
+        fabricio.run = functools.partial(fabricio.local, capture=True)
+
+        # uncomment row below to skip file uploading (e.g. docker-compose.yml)
+        # fab.put = lambda *args, **kwargs: None
+
     fab.env.update(
         roledefs={
             'web': ['localhost'],