+++ changelog.rst
@@ -4,6 +4,7 @@ Changelog
 Release 0.5.2
 -------------
 
+- Change: ``fabricio.infrastructure``: removed ``name`` parameter
 - Fix: full unicode support for Python 2 (`#100`_)
 
 .. _#100: https://github.com/renskiy/fabricio/issues/100